fixed all the remaining functions

+++ MusicMazter/MusicMazter/Inventory.cs
@@ -4,10 +4,10 @@ namespace MusicMazter
     public class Inventory
     {
         public string Name { get; set; }
-        public double Price { get; set; }
+        public int Price { get; set; }
         public int Quantity { get; set; }
 
-        public Inventory(string name, double price, int quantity)
+        public Inventory(string name, int price, int quantity)
         {
             Name = name;
             Price = price;
